[test] Revert to0e1adfbebb86780ef7dbae93b722febf12f6d4e7

+++ cpp_src/net/listener.h
@@ -61,7 +61,7 @@ protected:
 	ev::async async_;
 	std::shared_ptr<Shared> shared_;
 	vector<std::unique_ptr<IServerConnection>> connections_;
-	uint64_t id_;
+	int id_;
 };
 }  // namespace net
 }  // namespace reindexer
